requirements.txt
jupyter
pandas
lxml
bson
tqdm
urllib3
pytools

# higher versions require Python 3.7
numpy==1.19.0

beautifulsoup4

pyjnius

torch
torchtext

transformers
sentence-transformers
sentencepiece

thrift==0.13.0
spacy==2.2.3
